increase input size

+++ samples/sha512/sw/include/sha512.h
@@ -9,4 +9,4 @@
 uint64_t rotr(uint64_t input, uint8_t amnt);
 void doRound(uint64_t* input, uint8_t roundNumber, uint64_t word);
 void getwtschedule(uint64_t *m, uint64_t *schedule);
-void calculateHash(uint16_t blockcount, uint64_t* inputstring, uint64_t* buffers);
+void calculateHash(uint64_t blockcount, uint64_t* inputstring, uint64_t* buffers);
